% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_time_series_ops.r
\name{long2WASA_P}
\alias{long2WASA_P}
\title{writes WASA input file P}
\usage{
long2WASA_P(long, address)
}
\arguments{
\item{long}{is a long data frame object whose points are the centroids of the subbasins}

\item{address}{is the path where the file should be written}
}
\description{
writes WASA input file P
}
