% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_time_series_ops.r
\name{long2WASA_T}
\alias{long2WASA_T}
\title{writes WASA input file T}
\usage{
long2WASA_T(long, address)
}
\arguments{
\item{address}{is the path where the file should be written}

\item{stObj}{is a spacetime object whose points are the centroids of the subbasins}
}
\description{
writes WASA input file T
}
