% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_order_ops.R
\name{updown2idup}
\alias{updown2idup}
\title{returns a data frame with columns id for each subbasin and upstream.}
\usage{
updown2idup(df)
}
\arguments{
\item{df}{is a data frame imported from wasa with columns upstream and downstream}
}
\description{
returns a data frame with columns id for each subbasin and upstream.
}
