% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_order_ops.R
\name{updown2strahler}
\alias{updown2strahler}
\title{attributes strahler order to each basins after reading the routing table from wasa}
\usage{
updown2strahler(df)
}
\arguments{
\item{df}{is a data frame imported from wasa with columns upstream and downstream}
}
\description{
attributes strahler order to each basins after reading the routing table from wasa
}
